requirements.txt
python-telegram-bot==13.7
requests
urllib3==1.26.18
pymongo